+++ x-pack/test/upgrade/apps/maps/maps_smoke_tests.ts
@@ -16,6 +16,7 @@ export default function ({
   updateBaselines,
 }: FtrProviderContext & { updateBaselines: boolean }) {
   const PageObjects = getPageObjects(['common', 'maps', 'header', 'home', 'timePicker']);
+  const mapsHelper = getService('mapsHelper');
   const screenshot = getService('screenshots');
   const testSubjects = getService('testSubjects');
   const kibanaServer = getService('kibanaServer');
@@ -111,7 +112,7 @@ export default function ({
           );
           await PageObjects.header.waitUntilLoadingHasFinished();
           await PageObjects.maps.waitForLayersToLoad();
-          await PageObjects.maps.toggleLayerVisibilityRoadMap();
+          await mapsHelper.toggleLayerVisibilityRoadMap();
           await PageObjects.maps.toggleLayerVisibility('United Kingdom');
           await PageObjects.maps.toggleLayerVisibility('France');
           await PageObjects.maps.toggleLayerVisibility('United States');
@@ -141,7 +142,7 @@ export default function ({
           );
           await PageObjects.header.waitUntilLoadingHasFinished();
           await PageObjects.maps.waitForLayersToLoad();
-          await PageObjects.maps.toggleLayerVisibilityRoadMap();
+          await mapsHelper.toggleLayerVisibilityRoadMap();
           await PageObjects.timePicker.setCommonlyUsedTime('sample_data range');
           await PageObjects.maps.enterFullScreen();
           await PageObjects.maps.closeLegend();
@@ -167,8 +168,8 @@ export default function ({
           );
           await PageObjects.header.waitUntilLoadingHasFinished();
           await PageObjects.maps.waitForLayersToLoad();
-          await PageObjects.maps.toggleLayerVisibilityRoadMap();
-          await PageObjects.maps.toggleLayerVisibilityTotalRequests();
+          await mapsHelper.toggleLayerVisibilityRoadMap();
+          await mapsHelper.toggleLayerVisibilityTotalRequests();
           await PageObjects.timePicker.setCommonlyUsedTime('sample_data range');
           await PageObjects.maps.enterFullScreen();
           await PageObjects.maps.closeLegend();
