+++ x-pack/test/upgrade/config.ts
@@ -8,6 +8,7 @@
 import { FtrConfigProviderContext } from '@kbn/test';
 import { pageObjects } from './page_objects';
 import { ReportingAPIProvider } from './reporting_services';
+import { MapsHelper } from './maps_services';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
   const apiConfig = await readConfigFile(require.resolve('../api_integration/config'));
@@ -29,10 +30,11 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
       ...apiConfig.get('services'),
       ...functionalConfig.get('services'),
       reportingAPI: ReportingAPIProvider,
+      mapsHelper: MapsHelper,
     },
 
     junit: {
-      reportName: 'Upgrade Tests',
+      reportName: 'Kibana Core Tests',
     },
 
     timeouts: {
